Create `ocid` crate
Defines "Ocean Content IDs", which are self-describing values for
addressing content within Ocean.

Version 0 is the initial definition of this ID. Hopefully, it'll also be
the last. It is represented by `OcidV0`, which is made up of 2 version
bytes, 8 file size bytes, and 32 hash bytes.

The version 0 hash is derived from the BLAKE3 algorithm. This was chosen
for of its:
- Security against length extension, which SHA-2 is susceptible to
- Capability of verified streaming and incremental updates
- Performance